Compare references with results provided by current build.

+++ sources/force-render-png.sh
@@ -4,18 +4,25 @@ set -x
 
 SCRIPT_DIR=$(cd `dirname "$0"`; pwd)
 
-rm -rf ${SCRIPT_DIR}/../references/*
+THRESHOLD=5000
+MODE=$1
+PASS=0
+FAIL=0
+FAILED_TEST=""
+SYNFIG_OPTION=""
 
-
+if [ "$MODE" = "references" ]; then
+	rm -rf ${SCRIPT_DIR}/../references/*
+fi
 
 DEFAULT_VERSION=`cat ${SCRIPT_DIR}/default-version.txt`
 
 # Get synfig pre-build for default version
 
-$SYNFIG=""
+SYNFIG=""
 
 get-synfig-tar () {
-SYNFIG="/tmp/synfig-$VERSION/synfig"
+SYNFIG="/tmp/synfig-$VERSION/synfig $SYNFIG_OPTION"
 if [ ! -d "/tmp/synfig-$VERSION" ]; then
 	wget --quiet "https://sourceforge.net/projects/synfig/files/releases/$VERSION/linux/synfigstudio-$VERSION.x86_64.tar.bz2/download" -O "/tmp/synfig-$VERSION.tar.bz2"
 	mkdir -p /tmp/synfig-$VERSION
@@ -24,7 +31,7 @@ fi
 }
 
 get-synfig-appimage () {
-SYNFIG="/tmp/synfig-$VERSION.appimage --appimage-exec synfig" 
+SYNFIG="/tmp/synfig-$VERSION.appimage --appimage-exec synfig $SYNFIG_OPTION" 
 if [ ! -d "/tmp/synfig-$VERSION.appimage" ]; then
 	if [ "$#" -eq 3 ]; then
 		LINK=$3
@@ -40,7 +47,9 @@ get-synfig () {
 VERSION=$1
 PARSED_VERSION=${VERSION//./}
 
-if [ $PARSED_VERSION -lt 120 ]; then
+if [ "$MODE" = "results" ]; then 
+	SYNFIG="$HOME/synfig/bin/synfig"
+elif [ $PARSED_VERSION -lt 120 ]; then
 	get-synfig-tar $VERSION
 elif [ $PARSED_VERSION -eq 120 ]; then
 	get-synfig-appimage $VERSION "1.2.0-64bit-r2"
@@ -52,6 +61,10 @@ fi
 }
 
 COMPONENTS="blend-methods converters layers"
+
+if [ $MODE = "results" ]; then
+	mkdir -p ${SCRIPT_DIR}/../$MODE
+fi
 
 for COMPONENT in $COMPONENTS; do
 	pushd "${SCRIPT_DIR}/$COMPONENT"
@@ -70,14 +83,29 @@ for COMPONENT in $COMPONENTS; do
 			if [ ! -d "/tmp/synfig-$VERSION" ]; then
 				get-synfig $VERSION
 			fi
-			pushd ../../../references
+			pushd ../../../$MODE
 			mkdir -p ./$COMPONENT/$dir
 			popd
 			CURRENT_DIR=`pwd`
 			for sample in * ; do
 				# Renders every sif file present 
-				if [ "${sample##*.}" = "sif" ]; then
-					$SYNFIG --time 0 -i $CURRENT_DIR/$sample -o $CURRENT_DIR/../../../references/$COMPONENT/$dir/"${sample%.*}".png
+				NAME=${sample%.*}
+				EXT=${sample##*.}
+				if [ "$EXT" = "sif" ]; then
+					$SYNFIG --time 0 -i $CURRENT_DIR/$sample -o $CURRENT_DIR/../../../$MODE/$COMPONENT/$dir/"$NAME".png &> /dev/null
+				fi
+				if [ "$MODE" = "results" ]; then
+					TEST=$(compare -metric RMSE $CURRENT_DIR/../../../$MODE/$COMPONENT/$dir/"$NAME".png  $CURRENT_DIR/../../../references/$COMPONENT/$dir/"$NAME".png NULL 2>&1)
+					TEST=${TEST% *}
+					TEST=${TEST%.*}
+					if [ $TEST -lt $THRESHOLD ]; then
+						echo "$NAME passed"
+						PASS=$((PASS+1))
+					else
+						echo "SNAME failed"
+						FAIL=$((FAIL+1))
+						FAILED_TEST="$FAILED_TEST $NAME \n"
+					fi
 				fi
 			done
 			popd
@@ -85,3 +113,11 @@ for COMPONENT in $COMPONENTS; do
 	done
 	popd 
 done
+
+if [ "$MODE" = "results" ]; then
+	echo "TOTAL $((PASS+FAIL))"
+	echo "PASSED $PASS"
+	echo "FAILED $FAIL"
+	echo "FAILED TESTS"
+	printf "$FAILED_TEST"
+fi
